Move public functions for translation out of __init__
Given how Python handles module interfaces (i.e everything's
public), it's cleaner to define the interface in a submodule,
then import the public functions into __init__.
